refactor: Layered Architecture 적용
