+++ src/main/java/racingcar/dao/CarDao.java
@@ -5,5 +5,5 @@ import racingcar.dto.CarDto;
 import java.util.List;
 
 public interface CarDao {
-    void save(List<CarDto> carDtos);
+    void save(int gameId, List<CarDto> carDtos);
 }
