+++ src/main/java/racingcar/repository/RacingCarRepository.java
@@ -1,0 +1,8 @@
+package racingcar.repository;
+
+import racingcar.dto.RacingGameResultDto;
+
+public interface RacingCarRepository {
+
+    void save(RacingGameResultDto racingGameResultDto);
+}
