+++ src/main/java/racingcar/service/RacingCarService.java
@@ -1,0 +1,9 @@
+package racingcar.service;
+
+import racingcar.dto.request.RacingStartRequest;
+import racingcar.dto.response.RacingResultResponse;
+
+public interface RacingCarService {
+
+    RacingResultResponse play(RacingStartRequest racingStartRequest);
+}
