+++ src/main/java/racingcar/utils/ConnectionProvider.java
@@ -12,7 +12,6 @@ public class ConnectionProvider {
     public static Connection getConnection() {
         Connection connection = null;
         try {
-//            DriverManager.registerDriver(new org.h2.Driver());
             connection = DriverManager.getConnection("jdbc:h2:mem:testdb", "sa", "");
 
         } catch (SQLException e) {
